# Securix-Security-Dashboard/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Securix threat analysis and forensic case API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  ADMIN: 'ADMIN',
  INVESTIGATOR: 'INVESTIGATOR',
} as const;

export interface RegisterInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
  role: RegisterInputRole;
}

export interface AuthSession {
  token: string;
  username: string;
  role: string;
}

export interface User {
  id: number;
  username: string;
  role: string;
}

export interface DashboardStats {
  totalCases: number;
  maliciousCases: number;
  suspiciousCases: number;
  safeCases: number;
  openCases: number;
  averageRiskScore: number;
}

export interface ForensicCase {
  id: number;
  caseNumber: string;
  emailId: number;
  classification: string;
  riskScore: number;
  threatType: string;
  status: string;
  createdAt: string;
  updatedAt: string;
  summary: string;
}

export interface ThreatEvidence {
  id: number;
  evidenceId: string;
  caseId: string;
  year: string;
  evidenceType: string;
  evidenceValue: string;
  relatedThreat: string;
  location: string;
  forensicUse: string;
}

export interface CaseTimeline {
  id: number;
  caseId: number;
  eventTime: string;
  eventType: string;
  description: string;
  severity: string;
}

export type TimelineInputSeverity = typeof TimelineInputSeverity[keyof typeof TimelineInputSeverity];


export const TimelineInputSeverity = {
  INFO: 'INFO',
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
  CRITICAL: 'CRITICAL',
} as const;

export interface TimelineInput {
  /** @minLength 1 */
  eventType: string;
  /** @minLength 1 */
  description: string;
  severity: TimelineInputSeverity;
}

export interface Email {
  id: number;
  sender: string;
  recipient: string;
  subject: string;
  body: string;
}

export interface EmailInput {
  /** @minLength 1 */
  sender: string;
  /** @minLength 1 */
  recipient: string;
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  body: string;
}

export interface GeoLocationResult {
  ipAddress: string;
  country: string;
  state: string;
  city: string;
  isp: string;
}

export interface EmailAnalysisResult {
  emailId: number;
  classification: string;
  riskScore: number;
  explanation: string;
  urls: string[];
  domains: string[];
  ipAddresses: string[];
  emailAddresses: string[];
  geoLocations: GeoLocationResult[];
  matchedEvidence: ThreatEvidence[];
}

export interface CaseReport {
  caseDetails: ForensicCase;
  email: Email;
  evidence: ThreatEvidence[];
  timeline: CaseTimeline[];
}