# Securix-Security-Dashboard/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Securix threat analysis and forensic case API
servers:
  - url: /api
    description: Securix API base path
tags:
  - name: health
  - name: auth
  - name: dashboard
  - name: cases
  - name: evidence
  - name: email
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in to Securix
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a Securix user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "200":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid registration
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard risk statistics
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /cases:
    get:
      operationId: getCases
      tags: [cases]
      summary: List forensic cases
      responses:
        "200":
          description: Forensic cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ForensicCase"
  /cases/{caseId}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a forensic case
      parameters:
        - $ref: "#/components/parameters/CaseId"
      responses:
        "200":
          description: Forensic case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForensicCase"
        "404":
          description: Case not found
  /cases/{caseId}/timeline:
    get:
      operationId: getCaseTimeline
      tags: [cases]
      summary: Get a case timeline
      parameters:
        - $ref: "#/components/parameters/CaseId"
      responses:
        "200":
          description: Case timeline
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseTimeline"
    post:
      operationId: addCaseTimelineEvent
      tags: [cases]
      summary: Add a timeline event
      parameters:
        - $ref: "#/components/parameters/CaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimelineInput"
      responses:
        "200":
          description: Added timeline event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseTimeline"
  /cases/{caseId}/report:
    get:
      operationId: getCaseReport
      tags: [cases]
      summary: Generate a forensic case report
      parameters:
        - $ref: "#/components/parameters/CaseId"
      responses:
        "200":
          description: Case report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseReport"
  /cases/{caseId}/evidence:
    get:
      operationId: getCaseEvidence
      tags: [cases]
      summary: Get evidence matched to a case
      parameters:
        - $ref: "#/components/parameters/CaseId"
      responses:
        "200":
          description: Evidence records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ThreatEvidence"
  /evidence:
    get:
      operationId: getEvidence
      tags: [evidence]
      summary: List threat evidence
      responses:
        "200":
          description: Threat evidence
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ThreatEvidence"
  /email/analyze:
    post:
      operationId: analyzeEmail
      tags: [email]
      summary: Analyze an email and create a forensic case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailInput"
      responses:
        "200":
          description: Email analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailAnalysisResult"
components:
  parameters:
    CaseId:
      name: caseId
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 1 }
        password: { type: string, minLength: 1 }
    RegisterInput:
      type: object
      required: [username, password, role]
      properties:
        username: { type: string, minLength: 1 }
        password: { type: string, minLength: 1 }
        role: { type: string, enum: [ADMIN, INVESTIGATOR] }
    AuthSession:
      type: object
      required: [token, username, role]
      properties:
        token: { type: string }
        username: { type: string }
        role: { type: string }
    User:
      type: object
      required: [id, username, role]
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        role: { type: string }
    DashboardStats:
      type: object
      required:
        - totalCases
        - maliciousCases
        - suspiciousCases
        - safeCases
        - openCases
        - averageRiskScore
      properties:
        totalCases: { type: integer, format: int64 }
        maliciousCases: { type: integer, format: int64 }
        suspiciousCases: { type: integer, format: int64 }
        safeCases: { type: integer, format: int64 }
        openCases: { type: integer, format: int64 }
        averageRiskScore: { type: number, format: double }
    ForensicCase:
      type: object
      required:
        - id
        - caseNumber
        - emailId
        - classification
        - riskScore
        - threatType
        - status
        - createdAt
        - updatedAt
        - summary
      properties:
        id: { type: integer, format: int64 }
        caseNumber: { type: string }
        emailId: { type: integer, format: int64 }
        classification: { type: string }
        riskScore: { type: integer }
        threatType: { type: string }
        status: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        summary: { type: string }
    ThreatEvidence:
      type: object
      required:
        - id
        - evidenceId
        - caseId
        - year
        - evidenceType
        - evidenceValue
        - relatedThreat
        - location
        - forensicUse
      properties:
        id: { type: integer, format: int64 }
        evidenceId: { type: string }
        caseId: { type: string }
        year: { type: string }
        evidenceType: { type: string }
        evidenceValue: { type: string }
        relatedThreat: { type: string }
        location: { type: string }
        forensicUse: { type: string }
    CaseTimeline:
      type: object
      required: [id, caseId, eventTime, eventType, description, severity]
      properties:
        id: { type: integer, format: int64 }
        caseId: { type: integer, format: int64 }
        eventTime: { type: string, format: date-time }
        eventType: { type: string }
        description: { type: string }
        severity: { type: string }
    TimelineInput:
      type: object
      required: [eventType, description, severity]
      properties:
        eventType: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        severity: { type: string, enum: [INFO, LOW, MEDIUM, HIGH, CRITICAL] }
    Email:
      type: object
      required: [id, sender, recipient, subject, body]
      properties:
        id: { type: integer, format: int64 }
        sender: { type: string }
        recipient: { type: string }
        subject: { type: string }
        body: { type: string }
    EmailInput:
      type: object
      required: [sender, recipient, subject, body]
      properties:
        sender: { type: string, minLength: 1 }
        recipient: { type: string, minLength: 1 }
        subject: { type: string, minLength: 1 }
        body: { type: string, minLength: 1 }
    GeoLocationResult:
      type: object
      required: [ipAddress, country, state, city, isp]
      properties:
        ipAddress: { type: string }
        country: { type: string }
        state: { type: string }
        city: { type: string }
        isp: { type: string }
    EmailAnalysisResult:
      type: object
      required:
        - emailId
        - classification
        - riskScore
        - explanation
        - urls
        - domains
        - ipAddresses
        - emailAddresses
        - geoLocations
        - matchedEvidence
      properties:
        emailId: { type: integer, format: int64 }
        classification: { type: string }
        riskScore: { type: integer }
        explanation: { type: string }
        urls:
          type: array
          items: { type: string }
        domains:
          type: array
          items: { type: string }
        ipAddresses:
          type: array
          items: { type: string }
        emailAddresses:
          type: array
          items: { type: string }
        geoLocations:
          type: array
          items: { $ref: "#/components/schemas/GeoLocationResult" }
        matchedEvidence:
          type: array
          items: { $ref: "#/components/schemas/ThreatEvidence" }
    CaseReport:
      type: object
      required: [caseDetails, email, evidence, timeline]
      properties:
        caseDetails: { $ref: "#/components/schemas/ForensicCase" }
        email: { $ref: "#/components/schemas/Email" }
        evidence:
          type: array
          items: { $ref: "#/components/schemas/ThreatEvidence" }
        timeline:
          type: array
          items: { $ref: "#/components/schemas/CaseTimeline" }