# Securix-Security-Dashboard/lib/api-zod/src/generated/types/caseReport.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Securix threat analysis and forensic case API
 * OpenAPI spec version: 0.1.0
 */
import type { CaseTimeline } from './caseTimeline';
import type { Email } from './email';
import type { ForensicCase } from './forensicCase';
import type { ThreatEvidence } from './threatEvidence';

export interface CaseReport {
  caseDetails: ForensicCase;
  email: Email;
  evidence: ThreatEvidence[];
  timeline: CaseTimeline[];
}