# Securix-Security-Dashboard/lib/api-zod/src/generated/types/caseTimeline.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Securix threat analysis and forensic case API
 * OpenAPI spec version: 0.1.0
 */

export interface CaseTimeline {
  id: bigint;
  caseId: bigint;
  eventTime: Date;
  eventType: string;
  description: string;
  severity: string;
}