# Securix-Security-Dashboard/lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Securix threat analysis and forensic case API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalCases: bigint;
  maliciousCases: bigint;
  suspiciousCases: bigint;
  safeCases: bigint;
  openCases: bigint;
  averageRiskScore: number;
}