# Securix-Security-Dashboard/lib/api-zod/src/generated/types/emailAnalysisResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Securix threat analysis and forensic case API
 * OpenAPI spec version: 0.1.0
 */
import type { GeoLocationResult } from './geoLocationResult';
import type { ThreatEvidence } from './threatEvidence';

export interface EmailAnalysisResult {
  emailId: bigint;
  classification: string;
  riskScore: number;
  explanation: string;
  urls: string[];
  domains: string[];
  ipAddresses: string[];
  emailAddresses: string[];
  geoLocations: GeoLocationResult[];
  matchedEvidence: ThreatEvidence[];
}