# Securix-Security-Dashboard/lib/api-zod/src/generated/types/emailInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Securix threat analysis and forensic case API
 * OpenAPI spec version: 0.1.0
 */

export interface EmailInput {
  /** @minLength 1 */
  sender: string;
  /** @minLength 1 */
  recipient: string;
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  body: string;
}