# Securix-Security-Dashboard/lib/api-zod/src/generated/types/forensicCase.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Securix threat analysis and forensic case API
 * OpenAPI spec version: 0.1.0
 */

export interface ForensicCase {
  id: bigint;
  caseNumber: string;
  emailId: bigint;
  classification: string;
  riskScore: number;
  threatType: string;
  status: string;
  createdAt: Date;
  updatedAt: Date;
  summary: string;
}