# Securix-Security-Dashboard/lib/api-zod/src/generated/types/geoLocationResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Securix threat analysis and forensic case API
 * OpenAPI spec version: 0.1.0
 */

export interface GeoLocationResult {
  ipAddress: string;
  country: string;
  state: string;
  city: string;
  isp: string;
}