# Securix-Security-Dashboard/lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Securix threat analysis and forensic case API
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterInputRole } from './registerInputRole';

export interface RegisterInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
  role: RegisterInputRole;
}