# Securix-Security-Dashboard/lib/api-zod/src/generated/types/threatEvidence.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Securix threat analysis and forensic case API
 * OpenAPI spec version: 0.1.0
 */

export interface ThreatEvidence {
  id: bigint;
  evidenceId: string;
  caseId: string;
  year: string;
  evidenceType: string;
  evidenceValue: string;
  relatedThreat: string;
  location: string;
  forensicUse: string;
}