# Securix-Security-Dashboard/lib/api-zod/src/generated/types/timelineInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Securix threat analysis and forensic case API
 * OpenAPI spec version: 0.1.0
 */
import type { TimelineInputSeverity } from './timelineInputSeverity';

export interface TimelineInput {
  /** @minLength 1 */
  eventType: string;
  /** @minLength 1 */
  description: string;
  severity: TimelineInputSeverity;
}